Move fix_exception into data_gen_exceptions module
