+++ cumulusci/tasks/bulkdata/data_generation/data_generator_runtime.py
@@ -10,7 +10,7 @@ from jinja2 import nativetypes
 from cumulusci.core.template_utils import FakerTemplateLibrary, faker_template_library
 
 from .template_funcs import template_funcs
-from .data_gen_exceptions import DataGenError, DataGenSyntaxError
+from .data_gen_exceptions import DataGenSyntaxError
 
 
 # Runtime objects and algorithms used during the generation of rows.
@@ -182,19 +182,6 @@ def evaluate_function(func, args: Sequence, kwargs: Mapping, context):
         }
     value = func(*args, **kwargs)
     return value
-
-
-def fix_exception(message, parentobj, e):
-    """Add filename and linenumber to an exception if needed"""
-    filename, line_num = parentobj.filename, parentobj.line_num
-    if isinstance(e, DataGenError):
-        if not e.filename:
-            e.filename = filename
-        if not e.line_num:
-            e.line_num = line_num
-        raise e
-    else:
-        raise DataGenError(message, filename, line_num) from e
 
 
 class ObjectRow:
